requirements-dev.txt
-r requirements.txt
pytest==8.3.5
pytest-flask==1.3.0
pytest-cov==6.1.1